requirements.txt
PyGithub
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
PyYAML
